Issue #62: Refactoring of GitHubServiceSpi to delete a single webhook, not all Webhooks in a repo

+++ services/github-service-impl/src/main/java/org/kontinuity/catapult/service/github/impl/kohsuke/KohsukeGitHubServiceImpl.java
@@ -30,7 +30,9 @@ import org.kontinuity.catapult.service.github.spi.GitHubServiceSpi;
  */
 final class KohsukeGitHubServiceImpl implements GitHubService, GitHubServiceSpi {
 
-    private static final String GITHUB_WEBHOOK_WEB = "web";
+    private static final String WEBHOOK_URL = "url";
+
+	private static final String GITHUB_WEBHOOK_WEB = "web";
 
     private static final Logger log = Logger.getLogger(KohsukeGitHubServiceImpl.class.getName());
 
@@ -176,7 +178,7 @@ final class KohsukeGitHubServiceImpl implements GitHubService, GitHubServiceSpi 
             throw new RuntimeException(ioe);
         }
         Map<String, String> configuration = new HashMap<>();
-    	configuration.put("url", webhookUrl.toString());
+    	configuration.put(WEBHOOK_URL, webhookUrl.toString());
     	configuration.put("content_type", "json");
 
         List<GHEvent> githubEvents = Stream.of(events).map(event -> GHEvent.valueOf(event.name())).collect(Collectors.toList());
@@ -199,7 +201,7 @@ final class KohsukeGitHubServiceImpl implements GitHubService, GitHubServiceSpi 
     /**
      * {@inheritDoc}
      */
-    public void deleteWebhooks(final GitHubRepository repository) throws IllegalArgumentException {
+    public void deleteWebhook(final GitHubRepository repository, GitHubWebhook webhook) throws IllegalArgumentException {
     	if(repository == null) {
     		throw new IllegalArgumentException("repository must be specified");
     	}
@@ -208,7 +210,10 @@ final class KohsukeGitHubServiceImpl implements GitHubService, GitHubServiceSpi 
     		repo = delegate.getRepository(repository.getFullName());
     		
     		for (GHHook hook: repo.getHooks()) {
-        		hook.delete();
+			if(hook.getConfig().get(WEBHOOK_URL).equals(webhook.getUrl())) {
+				hook.delete();
+				break;
+			}
         	}
         } catch (final IOException ioe) {
             // Check for repo not found (this is how Kohsuke Java Client reports the error)
